+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # MountsJournal
 
-* Fixed "Use Underlight Angler" in combat.
+* Added the ability to hide mounts.